Add a more friendly error message for Rails3
If no database.mongo.yml file is defined and the user hasnt set up stuff manually, print a better error message to the console. I thought about using Rails.logger.error for this, however it doesn't show up in `rails console` which is primarily where people will be seeing it I think. So using plain old `puts`, it should show up in both the logs and in rails console