Hooking in the safety DSL on a per-query basis.
The safety DSL can now be used like so:

Person.safely.create(:title => "Sir")
person.safely.save!

The Safety proxy delegates using method_missing
except in the cases of methods that take an
attributes hash. This is because we did not want
to throw a :safe => true in the hash and then
remove it later - safe is too common of a word
and could be used as a field/attribute and
collide with this functionality. I've deemed the
slight duplication acceptable.