correct usage of ActiveSupport::Concern
You should not use include inside an included block as this prevents
developers from extending any of the mongoid api in the correct way.
Module inclusion should build up the inheritance chain with each include,
so that the last included module is first in the chain. Including in an
included block kills this by insuring your module is included last.

Check out the source for ActiveSupport::Concern to see what I mean.