Allow reordering of foreign keys even with pulls/pushes
When there were pushes (or pulls) as well as elements being reordered,
save would only push (or pull) elements. The changed order of the other
elements were lost.

With this change, checking is done to ensure that we only push when the
new elements are at the end, and pull when the remaining element's order
has not changed.