Handle batch state passing in M/R aggregations
Mongo batches reduce jobs in groups of 100 and passes the state along
via the first document each time in the reduce function. This was
causing the count in the aggregations to reset every 100 documents that
were passed through the reduce. The effect would throw off any
calculations for avg and count with result sets larger than 100 docs.

[ fix #2482 ]