Remove full path atomic selectors.
Since SERVER-831 has been open with MongoDB for 2 years, 8 months with
no resolution in sight, it's pretty much safe to say it won't be getting
fix and we don't need the code to support it anymore.

With this removal, when updating embedded documents the selector now
used is now much simpler - only { _id: root.id } instead of being smart
to location the children as well.

This now allows embedded documents that had been persisted outside
Mongoid to be loaded and assigned ids from within Mongoid, and properly
saved.

[ fix #2496 ]